smart-sqlite-agent/requirements.txt
fastmcp
uvicorn
llama-index
llama-index-tools-mcp
python-dotenv
openai